Refactor code to add cli
